+++ src/umpteen/repr.rs
@@ -1,5 +1,6 @@
 pub mod ast;
 pub mod bytecode;
+pub mod fnc;
 pub mod object;
 pub mod token;
 pub mod value;
